#S06P31D204-93 feat: 일대다 연관관계 추가 및 수정

+++ BE/src/main/java/com/marketganada/db/repository/AddressBookRepository.java
@@ -3,5 +3,5 @@ package com.marketganada.db.repository;
 import com.marketganada.db.entity.AddressBook;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AddressBookRepository extends JpaRepository<AddressBook,Integer> {
+public interface AddressBookRepository extends JpaRepository<AddressBook,Long> {
 }
