+++ BE/src/main/java/com/marketganada/db/repository/AuctionImgRepository.java
@@ -3,5 +3,5 @@ package com.marketganada.db.repository;
 import com.marketganada.db.entity.AuctionImg;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuctionImgRepository extends JpaRepository<AuctionImg,Integer> {
+public interface AuctionImgRepository extends JpaRepository<AuctionImg,Long> {
 }
