+++ BE/src/main/java/com/marketganada/db/repository/AuctionRepository.java
@@ -3,5 +3,5 @@ package com.marketganada.db.repository;
 import com.marketganada.db.entity.Auction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuctionRepository extends JpaRepository<Auction,Integer> {
+public interface AuctionRepository extends JpaRepository<Auction,Long> {
 }
