+++ BE/src/main/java/com/marketganada/db/repository/CategoryLargeRepository.java
@@ -3,5 +3,5 @@ package com.marketganada.db.repository;
 import com.marketganada.db.entity.CategoryLarge;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryLargeRepository extends JpaRepository<CategoryLarge,Integer> {
+public interface CategoryLargeRepository extends JpaRepository<CategoryLarge,Long> {
 }
