+++ BE/src/main/java/com/marketganada/db/repository/CategorySmallRepository.java
@@ -3,5 +3,5 @@ package com.marketganada.db.repository;
 import com.marketganada.db.entity.CategorySmall;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategorySmallRepository extends JpaRepository<CategorySmall,Integer> {
+public interface CategorySmallRepository extends JpaRepository<CategorySmall,Long> {
 }
