+++ BE/src/main/java/com/marketganada/db/repository/ProductHistoryRepository.java
@@ -3,5 +3,5 @@ package com.marketganada.db.repository;
 import com.marketganada.db.entity.ProductHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductHistoryRepository extends JpaRepository<ProductHistory,Integer> {
+public interface ProductHistoryRepository extends JpaRepository<ProductHistory,Long> {
 }
